requirements.txt
# Grounded Industrial Defect Explanation — minimal implementation
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.24.0
PyYAML>=6.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
scipy>=1.11.0
opencv-python-headless>=4.8.0
scikit-image>=0.22.0
